refactor: replace dedicated onboarding page with modal-based approach
- Remove CheckOnboarding middleware and its test
- Delete standalone Onboarding.vue page component
- Add new OnboardingModal.vue component for inline onboarding
- Update MainV2.vue to integrate onboarding modal
- Remove onboarding route from web.php
- Simplify bootstrap/app.php by removing onboarding middleware

This change improves UX by keeping users in context during onboarding
rather than redirecting to a separate page.
