+++ routes/web.php
@@ -14,11 +14,6 @@ Route::get('/', function () {
     return Inertia::render('Welcome');
 })->name('home');
 
-// Onboarding Route
-Route::get('/onboarding', function () {
-    return Inertia::render('Onboarding');
-})->name('onboarding');
-
 Route::get('dashboard', function () {
     return Inertia::render('Dashboard');
 })->name('dashboard');
